backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
starlette
itsdangerous
google-auth
google-auth-oauthlib
google-api-python-client
requests
httpx
openai
pdfplumber
python-docx
openpyxl
pytesseract
pillow
beautifulsoup4
groq